pyproject.toml
[project]
name = "at_relay"
version = "0.1.0"
authors = [{ name = "Nicolai Buchwitz", email = "nb@tipi-net.de" }]
description = "AT command controlled relay interface"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.7"
keywords = ["testing", "hardware"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["pyserial"]

[project.optional-dependencies]
"test" = ["pytest", "mock_serial", "pytest-cov", "ruff", "black"]

[project.urls]
# documentation = ""
repository = "https://github.com/nbuchwitz/at_relay/"

[build-system]
requires = ["setuptools>=64",]
build-backend = "setuptools.build_meta"

[project.scripts]
eol-frontend = "frontend.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["at_relay"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/at_relay", "tests"]

[tool.ruff]
line-length = 100
select = [
    "ANN", # flake8-annotations
    "B",   # flake8-bugbear
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # Pyflakes
    "Q",   # flake8-quotes
]
ignore = [
    "ANN101", # missing-type-self
    "D107",   # undocumented-public-init
]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.mccabe]
max-complexity = 10
